% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_com.R
\name{dataframe_com}
\alias{dataframe_com}
\title{Commodity Dataframe}
\usage{
dataframe_com(dataset, commodityname)
}
\arguments{
\item{dataset}{It is the dataset of keymaterials}

\item{commodityname}{the name of commodity such as Aluminum, etc.}
}
\value{
a price range of the commodity
}
\description{
Commodity Dataframe
}
